add Rakefile & .travis.yml, rename Gem, update version
