+++ .travis.yml
@@ -1,0 +1,12 @@
+branches:
+  only:
+    - master
+language: ruby
+rvm:
+- 2.1.0
+- 2.0.0
+- 1.9.3
+- 2.1.1
+notifications:
+  email:
+  - t+travisci@heckman.io
