+++ lib/lita_forecast.rb
