+++ lib/lita_forecast/version.rb
@@ -3,5 +3,5 @@
 # LitaForcast module, just for version handling
 #
 module LitaForecast
-  VERSION = '0.0.1'
+  VERSION = '0.0.2'
 end
